Imguiのサイズと中身の追加、その他コメントアウト等の追加

+++ AL3/SourceCodes/3D/FbxObject3D.cpp
@@ -347,9 +347,9 @@ void FbxObject3D::Update(bool isShadowCamera)
 		//matRot.r[0].m128_f32[0] = fbxMatRot.mData[0] / 180.0f * 3.14f;
 		//matRot.r[1].m128_f32[1] = fbxMatRot.mData[1] / 180.0f * 3.14f;
 		//matRot.r[2].m128_f32[2] = fbxMatRot.mData[2] / 180.0f * 3.14f;
-		matRot *= DirectX::XMMatrixRotationZ(fbxMatRot.mData[0] / 180.0f * 3.14f);
-		matRot *= DirectX::XMMatrixRotationX(fbxMatRot.mData[1] / 180.0f * 3.14f);
-		matRot *= DirectX::XMMatrixRotationY(fbxMatRot.mData[2] / 180.0f * 3.14f);
+		matRot *= DirectX::XMMatrixRotationZ(static_cast<float>(fbxMatRot.mData[0]) / 180.0f * 3.14f);
+		matRot *= DirectX::XMMatrixRotationX(static_cast<float>(fbxMatRot.mData[1]) / 180.0f * 3.14f);
+		matRot *= DirectX::XMMatrixRotationY(static_cast<float>(fbxMatRot.mData[2]) / 180.0f * 3.14f);
 		localMatRots.push_back(matRot);
 
 		//éËÇÃÉèÅ[ÉãÉhçsóÒéÊìæ
